fix(watch): implement watch functionality using Rspack
